requirements.txt
pytorch
transformers
numpy
tqdm